requirements.txt
requests>=2.28.0
pandas>=1.5.0
numpy>=1.23.0
sqlalchemy>=2.0.0
pymysql>=1.0.0
xgboost>=1.7.0
scikit-learn>=1.2.0
matplotlib>=3.6.0
seaborn>=0.12.0
flask>=2.3.0
tabpy-client>=0.9.0
tabpy-server>=0.9.0
python-dotenv>=1.0.0
schedule>=1.2.0
jupyter>=1.0.0
notebook>=6.5.0
